requirements.txt
openai>=1.59.7
requests>=2.31.0
python-dotenv==1.0.1
pytest>=7.4.0
flake8>=3.9.2
pylint>=2.17.4
black>=23.3.0
typing-extensions>=4.7.1
anthropic
pinecone
pymysql>=1.0.2
DBUtils>=3.0.2
typing>=3.7.4
websockets
selenium
webdriver_manager
markdown
azure.storage.blob
aipolabs
